editorconfig: Add EditorConfig configuration
As this is an open source project, It is essential to maintain a consistent
coding style for code-readability, EditorConfig here automatically adjusts settings
like indentation, line endings, end of lines, etc. when editing a file as per the
configuration when using a supported editor or installing a plugin.

For more info:
[1] https://editorconfig.org
[2] https://www.topbug.net/blog/2012/03/14/use-editorconfig-to-maintain-consistent-coding-styles-between-different-editors-and-ides/

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>